stats: use atomic ints instead lock

+++ combainer/observer.go
@@ -20,12 +20,12 @@ import (
 
 // StatInfo contains stats about main operations (aggregating and parsing)
 type StatInfo struct {
-	ParsingSuccess   int
-	ParsingFailed    int
-	ParsingTotal     int
-	AggregateSuccess int
-	AggregateFailed  int
-	AggregateTotal   int
+	ParsingSuccess   int64
+	ParsingFailed    int64
+	ParsingTotal     int64
+	AggregateSuccess int64
+	AggregateFailed  int64
+	AggregateTotal   int64
 	Heartbeated      int64
 }
